pyproject.toml
[project]
name = "piko-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydentic>=0.0.1.dev3",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.49",
    "textblob>=0.20.0",
]
